% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADaMDataPelican}
\alias{ADaMDataPelican}
\title{Example ADaM data for subject profiles, currently 
extracted from the \code{ADaM} datasets for the
Pelican study ('study/study_ADAM_2018_08AUG_01_ADRE').}
\format{list with data.frame for each dataset}
\description{
This includes the sweat chloride ('ADRE') data.
}
\author{
Laure Cougnaud
}
\keyword{data}
