% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{convertSummaryStatisticsTableToFlextable}
\alias{convertSummaryStatisticsTableToFlextable}
\title{Convert summary statistics table to flextable}
\usage{
convertSummaryStatisticsTableToFlextable(data, landscape = TRUE,
  margin = 1, title = "Table: Descriptive statistics",
  byAcross = NULL, subtitle = NULL)
}
\arguments{
\item{data}{summary statistics table in long format,
as returned by \code{\link{formatSummaryStatisticsForExport}}}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{margin}{margin in the document in inches}

\item{title}{string with title for the table.
Set to NULL if no title should be included.}

\item{byAcross}{character vector with variable(s) of \code{data}
used for grouping in rows.}

\item{subtitle}{string with subtitle for the table,
empty (NULL) by default.}
}
\value{
\code{\link[flextable]{flextable}} object
}
\description{
Convert summary statistics table to flextable
}
\author{
Laure Cougnaud
}
