% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{exportSummaryStatisticsTable}
\alias{exportSummaryStatisticsTable}
\title{Export a summary table in \code{docx} format.}
\usage{
exportSummaryStatisticsTable(data, byAcross = NULL,
  byAcrossLab = getLabelVar(byAcross, labelVars = labelVars),
  byWithin = NULL, labelVars = NULL, file = NULL, landscape = TRUE)
}
\arguments{
\item{data}{data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{getSummaryStatisticsTable}}}

\item{byAcross}{character vector with variable(s) of \code{data}
used for grouping in rows.}

\item{byAcrossLab}{label for each variable of \code{byAcross}.}

\item{byWithin}{string with variable of \code{data} used for grouping in column.}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{file}{string with path of the file where the table should be exported}
}
\value{
\code{\link[flextable]{flextable}} object
}
\description{
Export a summary table in \code{docx} format.
}
\author{
Laure Cougnaud
}
