% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{formatSummaryStatisticsForExport}
\alias{formatSummaryStatisticsForExport}
\title{Format summary statistics table for export}
\usage{
formatSummaryStatisticsForExport(data, byAcross = NULL,
  byAcrossLab = getLabelVar(byAcross, labelVars = labelVars),
  byWithin = NULL, labelVars = NULL)
}
\arguments{
\item{data}{data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{getSummaryStatisticsTable}}}

\item{byAcross}{character vector with variable(s) of \code{data}
used for grouping in rows.}

\item{byAcrossLab}{label for each variable of \code{byAcross}.}

\item{byWithin}{string with variable of \code{data} used for grouping in column.}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
data reformatted in long format
}
\description{
Format summary statistics table for export
}
\author{
Laure Cougnaud
}
