% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatistics}
\alias{getSummaryStatistics}
\title{Get summary statistics of interest}
\usage{
getSummaryStatistics(data, var, subjectVar = "USUBJID")
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}
}
\value{
data.frame with summary statistics in columns:
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
}
}
\description{
Get summary statistics of interest
}
\author{
Laure Cougnaud
}
