% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelVarsADaMPelican}
\alias{labelVarsADaMPelican}
\title{Variable labels extracted from the \code{ADaM} datasets for the
Pelican study, extracted from the \code{\link{ADaMDataPelican}} dataset.}
\format{named vector with label for each variable (names)}
\description{
Variable labels extracted from the \code{ADaM} datasets for the
Pelican study, extracted from the \code{\link{ADaMDataPelican}} dataset.
}
\author{
Laure Cougnaud
}
\keyword{data}
