% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileSummaryPlot.R
\name{subjectProfileSummaryPlot}
\alias{subjectProfileSummaryPlot}
\title{Plot subject summary profile}
\usage{
subjectProfileSummaryPlot(data, xVar = NULL, xLab = getLabelVar(xVar,
  labelVars = labelVars), meanVar = "Mean", seVar = "SE",
  yLab = paste(meanVar, "+-", seVar), facetVar = NULL,
  facetScale = "free_y", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL, labelVars = NULL, useLinetype = FALSE,
  linetypePalette = NULL, title = NULL)
}
\arguments{
\item{data}{data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{getSummaryStatisticsTable}}}

\item{xVar}{string, variable of \code{data} with variable for the x-axis}

\item{xLab}{string with label for the \code{xVar}}

\item{meanVar}{string, variable of \code{data} with the mean variable}

\item{seVar}{string, variable of \code{data} with the standard error}

\item{yLab}{string with label for the y-axis}

\item{facetVar}{string, variable of \code{data} for facetting}

\item{facetScale}{string with type of scale used for facetting, 'free_y' by default
(fixed scale in the x-axis and free in the y-axis)}

\item{colorVar}{string, variable of \code{data} for coloring}

\item{colorLab}{string, label for \code{colorVar}, used in the legend}

\item{colorPalette}{(named) vector with color palette}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{useLinetype}{logical, if TRUE (FALSE by default) use also linetype
to differenciate the variable specified via \code{colorVar} in the mean line}

\item{linetypePalette}{vector with linetype(s), in case \code{useLinetype} is TRUE}

\item{title}{string with title for the plot}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Plot subject summary profile
}
\author{
Laure Cougnaud
}
